refactor: 더알아보기 팝업 AllOptions로 옮김
