+++ front/src/Pages/CustomPage/OptionPage/index.jsx
@@ -4,48 +4,24 @@ import AllOptions from "./AllOptions";
 import Button from "@Components/Common/Button/Button";
 import useButtonNavigation from "@hooks/useButtonNavigation";
 import { options } from "./optionData";
-import useOnClickPopUp from "@hooks/useOnClickPopUp";
-import OptionPopup from "./OptionPopup";
-import OverlaidPopup from "@Components/Common/OverlaidPopup";
 
 const OptionPage = () => {
   // 선택한 옵션들 상태관리
   const [selectOption, setSelectOption] = useState([]);
   const handleSelectOption = (option) => {
-    if (hasOption(option))
+    if (checkOptionSelected(option))
       setSelectOption((prev) => prev.filter((opt) => opt !== option));
     else setSelectOption((prev) => [...prev, option]);
   };
-  const hasOption = (option) => {
+  const checkOptionSelected = (option) => {
     return selectOption.includes(option);
   };
 
   const [selectedTab, setSelectedTab] = useState("추가 옵션");
-  const optionPopupRef = useRef();
-  const { isPopupOpen, openPopup, closePopup } =
-    useOnClickPopUp(optionPopupRef);
-  const [popupOptionName, setPopupOptionName] = useState(null);
-  const handleOpenPopup = (data) => {
-    setPopupOptionName(data);
-    openPopup();
-  };
   const move = useButtonNavigation();
 
   return (
     <Wrapper>
-      {isPopupOpen && (
-        <OverlaidPopup
-          component={
-            <OptionPopup
-              popupRef={optionPopupRef}
-              closePopup={closePopup}
-              popupOptionName={popupOptionName}
-              handleSelectOption={handleSelectOption}
-              hasOption={hasOption}
-            />
-          }
-        />
-      )}
       <TabWrapper>
         <TabItem
           selected={selectedTab === "추가 옵션"}
@@ -63,9 +39,8 @@ const OptionPage = () => {
       <AllOptions
         tab={selectedTab}
         options={options}
-        handleOpenPopup={handleOpenPopup}
         handleSelectOption={handleSelectOption}
-        hasOption={hasOption}
+        checkOptionSelected={checkOptionSelected}
       />
       <ButtonContainer>
         <Button
